examples/python-fastapi/requirements.txt
requests
python-dotenv
fastapi[all]
pyyaml 
ua-parser
user-agents